extract the structured_storage related code into separate file
cleanup JNI calls, error checking and logging
wrap the changes to file.c in #ifdef ANDROID
