+++ android/app/src/main/cpp/mobileffmpeg.c
@@ -92,9 +92,6 @@ struct CallbackData *callbackDataTail;
 /** Global reference to the virtual machine running */
 static JavaVM *globalVm;
 
-/** Global application context */
-static jobject *appContext;
-
 /** Global reference of Config class in Java */
 static jclass configClass;
 
@@ -550,45 +547,6 @@ void *callbackThreadFunction() {
 }
 
 /**
- * Kudos to Harlan Chen, https://stackoverflow.com/a/46871051/192373
- */
-static jobject get_global_context(JNIEnv *env) {
-    jclass activityThread = (*env)->FindClass(env, "android/app/ActivityThread");
-    if ((*env)->ExceptionCheck(env)) {
-        (*env)->ExceptionClear(env);
-        LOGE("get_global_context FindClass %s exception", "android/app/ActivityThread");
-        return NULL;
-    }
-    jmethodID currentActivityThread = (*env)->GetStaticMethodID(env, activityThread, "currentActivityThread", "()Landroid/app/ActivityThread;");
-    if ((*env)->ExceptionCheck(env)) {
-        (*env)->ExceptionClear(env);
-        LOGE("get_global_context find %s exception", "currentActivityThread");
-        return NULL;
-    }
-    jobject at = (*env)->CallStaticObjectMethod(env, activityThread, currentActivityThread);
-    if ((*env)->ExceptionCheck(env)) {
-        (*env)->ExceptionClear(env);
-        LOGE("get_global_context %s %s exception", "android/app/ActivityThread", "currentActivityThread");
-        return NULL;
-    }
-    jmethodID getApplication = (*env)->GetMethodID(env, activityThread, "getApplication", "()Landroid/app/Application;");
-    if ((*env)->ExceptionCheck(env)) {
-        (*env)->ExceptionClear(env);
-        LOGE("get_global_context find %s exception", "getApplication");
-        return NULL;
-    }
-    jobject context = (*env)->CallObjectMethod(env, at, getApplication);
-    if ((*env)->ExceptionCheck(env)) {
-        (*env)->ExceptionClear(env);
-        LOGE("get_global_context %s exception", "getApplication");
-        return NULL;
-    }
-
-    return (*env)->NewGlobalRef(env, context);
-    // TODO: release local refs
-}
-
-/**
  * Called when 'mobileffmpeg' native library is loaded.
  *
  * @param vm pointer to the running virtual machine
@@ -653,227 +611,9 @@ jint JNI_OnLoad(JavaVM *vm, void *reserved) {
     monitorInit();
     logInit();
 
-    appContext = get_global_context(env);
-    LOGI("set appContext=%p", appContext);
-
     return JNI_VERSION_1_6;
 }
 
-#include <unistd.h>
-#include <stdlib.h>
-#include <fcntl.h>
-
-static jclass    android_net_Uri = NULL; /* global ref */
-
-static jmethodID android_net_Uri_parse = NULL;
-static jmethodID android_content_Context_getContentResolver = NULL;
-static jmethodID android_content_ContentResolver_openFileDescriptor = NULL;
-static jmethodID android_content_ContentResolver_query = NULL;
-static jmethodID android_os_ParcelFileDescriptor_getFd = NULL;
-static jmethodID android_database_Cursor_moveToFirst = NULL;
-static jmethodID android_database_Cursor_getString = NULL;
-static jmethodID android_database_Cursor_getColumnIndex = NULL;
-static jmethodID android_database_Cursor_close = NULL;
-
-static jstring DISPLAY_NAME = NULL;
-static jobject android_content_ContentResolver_instance = NULL;  /* global ref */
-
-#define JNI_CHECK_EXTRA
-#define JNI_CHECK(j_object, format, ...) { \
-    if ((*env)->ExceptionCheck(env)) { \
-        (*env)->ExceptionClear(env); \
-        LOGE("%s:%d JNI exception, cannot get " #j_object " from " format, __PRETTY_FUNCTION__, __LINE__, __VA_ARGS__); \
-        JNI_CHECK_EXTRA; \
-        (*env)->PopLocalFrame(env, NULL); \
-        return JNI_ERR; \
-    } \
-    if (!j_object) { \
-        LOGE("%s:%d cannot get" #j_object " from " format, __PRETTY_FUNCTION__, __LINE__, __VA_ARGS__); \
-        JNI_CHECK_EXTRA; \
-        (*env)->PopLocalFrame(env, NULL); \
-        return JNI_ERR; \
-    } \
-}
-
-static int get_static_method_id(JNIEnv *env, const char *class_name, const char *name, const char *signature, jclass *global_class_reference, jmethodID *method_id) {
-    jclass class_reference = (*env)->FindClass(env, class_name);
-    JNI_CHECK(class_reference, "%s", class_name);
-    *global_class_reference = (*env)->NewGlobalRef(env, class_reference);
-    JNI_CHECK(*global_class_reference, "%p %s", class_reference, class_name);
-    (*env)->DeleteLocalRef(env, class_reference);
-
-    *method_id = (*env)->GetStaticMethodID(env, *global_class_reference, name, signature);
-    JNI_CHECK(*method_id, "%p class %s static method %s for %s", *global_class_reference, class_name, name, signature);
-    return JNI_OK;
-}
-
-static int get_method_id(JNIEnv *env, const char *class_name, const char *name, const char *signature, jmethodID *method_id) {
-    jclass class_reference = (*env)->FindClass(env, class_name);
-    JNI_CHECK(class_reference, "%s", class_name);
-
-    *method_id = (*env)->GetMethodID(env, class_reference, name, signature);
-    JNI_CHECK(*method_id, "%p class %s method %s for %s", class_reference, class_name, name, signature);
-    (*env)->DeleteLocalRef(env, class_reference);
-    return JNI_OK;
-}
-
-int match_ext_from_content(const char *content, const char *extensions);
-int get_fd_from_content(const char *content, int access) {
-
-    int fd = -1;
-
-    JNIEnv *env;
-    int ret = (*globalVm)->GetEnv(globalVm, (void **)&env, JNI_VERSION_1_6);
-    if (!env || ret != JNI_OK) {
-        LOGE("%s:%d cannot get env", __PRETTY_FUNCTION__, __LINE__);
-        return -1;
-    }
-
-    ret = (*env)->PushLocalFrame(env, 10);
-
-    if (!android_net_Uri) {
-        if (get_static_method_id(env, "android/net/Uri", "parse", "(Ljava/lang/String;)Landroid/net/Uri;", &android_net_Uri, &android_net_Uri_parse) != JNI_OK)
-            return JNI_ERR;
-
-        if (get_method_id(env, "android/content/Context", "getContentResolver", "()Landroid/content/ContentResolver;", &android_content_Context_getContentResolver) != JNI_OK)
-            return JNI_ERR;
-
-        if (get_method_id(env, "android/content/ContentResolver", "openFileDescriptor", "(Landroid/net/Uri;Ljava/lang/String;)Landroid/os/ParcelFileDescriptor;", &android_content_ContentResolver_openFileDescriptor) != JNI_OK)
-            return JNI_ERR;
-
-        if (get_method_id(env, "android/content/ContentResolver", "query", "(Landroid/net/Uri;[Ljava/lang/String;Ljava/lang/String;[Ljava/lang/String;Ljava/lang/String;)Landroid/database/Cursor;", &android_content_ContentResolver_query) != JNI_OK)
-            return JNI_ERR;
-
-        if (get_method_id(env, "android/database/Cursor", "moveToFirst", "()Z", &android_database_Cursor_moveToFirst) != JNI_OK)
-            return JNI_ERR;
-
-        if (get_method_id(env, "android/database/Cursor", "getString", "(I)Ljava/lang/String;", &android_database_Cursor_getString) != JNI_OK)
-            return JNI_ERR;
-
-        if (get_method_id(env, "android/database/Cursor", "getColumnIndex", "(Ljava/lang/String;)I", &android_database_Cursor_getColumnIndex) != JNI_OK)
-            return JNI_ERR;
-
-        if (get_method_id(env, "android/database/Cursor", "close", "()V", &android_database_Cursor_close) != JNI_OK)
-            return JNI_ERR;
-
-        if (get_method_id(env, "android/os/ParcelFileDescriptor", "getFd", "()I", &android_os_ParcelFileDescriptor_getFd) != JNI_OK)
-            return JNI_ERR;
-
-        DISPLAY_NAME = (*env)->NewStringUTF(env, "_display_name"); // DocumentsContract.Document.COLUMN_DISPLAY_NAME
-        JNI_CHECK(DISPLAY_NAME, "DocumentsContract.Document.COLUMN_DISPLAY_NAME=%s", "_display_name");
-    }
-
-    const char *fmode = "r";
-    if (access & (O_WRONLY | O_RDWR)) {
-        fmode = "w";
-    }
-
-    LOGI("get_fd_from_content" " \"%s\" fd from %s", fmode, content);
-
-    jstring uriString = (*env)->NewStringUTF(env, content);
-    JNI_CHECK(uriString, "%s", content);
-
-    jstring fmodeString = (*env)->NewStringUTF(env, fmode);
-    JNI_CHECK(fmodeString, "%s", fmode);
-
-    jobject uri = (*env)->CallStaticObjectMethod(env, android_net_Uri, android_net_Uri_parse, uriString);
-    JNI_CHECK(uri, "%p", uriString);
-
-    if (android_content_ContentResolver_instance == NULL) {
-        jobject contentResolver = (*env)->CallObjectMethod(env, appContext, android_content_Context_getContentResolver);
-        JNI_CHECK(contentResolver, "%p", appContext);
-        android_content_ContentResolver_instance = (*env)->NewGlobalRef(env, contentResolver);
-        JNI_CHECK(android_content_ContentResolver_instance, "%p", contentResolver);
-    }
-
-    jobject parcelFileDescriptor = (*env)->CallObjectMethod(env, android_content_ContentResolver_instance, android_content_ContentResolver_openFileDescriptor, uri, fmodeString);
-    JNI_CHECK(parcelFileDescriptor, "%p (%s, %s)", android_content_ContentResolver_instance, content, fmode);
-
-    fd = (*env)->CallIntMethod(env, parcelFileDescriptor, android_os_ParcelFileDescriptor_getFd);
-    JNI_CHECK((fd >= 0), "%p", parcelFileDescriptor);
-
-    LOGI("match_ext_from_content: %d", match_ext_from_content(content, "mp4"));
-
-    (*env)->PopLocalFrame(env, NULL);
-    return dup(fd);
-}
-
-/* based on https://stackoverflow.com/a/25005243/192373
-
-  Cursor cursor = getContentResolver().query(uri, null, null, null, null);
-    try {
-      if (cursor != null && cursor.moveToFirst()) {
-        result = cursor.getString(cursor.getColumnIndex(DocumentsContract.Document.COLUMN_DISPLAY_NAME));
-      }
-    } finally {
-      cursor.close();
-    }
-*/
-
-static int get_filename_from_content(const char *content, char filename[]) {
-
-    if (!android_content_ContentResolver_instance) {
-        LOGE("%s:%d have not initialized android.content.ContentResolver instance", __PRETTY_FUNCTION__, __LINE__);
-        return JNI_ERR;
-    }
-
-    if (!android_content_ContentResolver_openFileDescriptor) {
-        LOGE("%s:%d have not initialized android.content.ContentResolver instance", __PRETTY_FUNCTION__, __LINE__);
-        return JNI_ERR;
-    }
-
-    JNIEnv *env;
-    int ret = (*globalVm)->GetEnv(globalVm, (void **)&env, JNI_VERSION_1_6);
-    if (!env || ret != JNI_OK) {
-        LOGE("%s:%d cannot get env", __PRETTY_FUNCTION__, __LINE__);
-        return JNI_ERR;
-    }
-
-    ret = (*env)->PushLocalFrame(env, 10);
-
-    jstring uriString = (*env)->NewStringUTF(env, content);
-    JNI_CHECK(uriString, "%s", content);
-
-    jobject uri = (*env)->CallStaticObjectMethod(env, android_net_Uri, android_net_Uri_parse, uriString);
-    JNI_CHECK(uri, "%p", uriString);
-
-    jobject cursor = (*env)->CallObjectMethod(env, android_content_ContentResolver_instance, android_content_ContentResolver_query, uri, NULL, NULL, NULL, NULL);
-    JNI_CHECK(cursor, "%p (%s, null, null, null, null)", android_content_ContentResolver_instance, content);
-
-#undef  JNI_CHECK_EXTRA
-#define JNI_CHECK_EXTRA (*env)->CallVoidMethod(env, cursor, android_database_Cursor_close)
-
-    jboolean move_res = (*env)->CallBooleanMethod(env, cursor, android_database_Cursor_moveToFirst);
-    JNI_CHECK(move_res, "%p moveToFirst()", cursor);
-
-    jint column_idx = (*env)->CallIntMethod(env, cursor, android_database_Cursor_getColumnIndex, DISPLAY_NAME);
-    JNI_CHECK((column_idx >= 0), "%p getColumnIndex(DISPLAY_NAME)", cursor);
-
-    jstring j_str = (jstring)(*env)->CallObjectMethod(env, cursor, android_database_Cursor_getString, column_idx);
-    JNI_CHECK(j_str, "%p getString (%d)", cursor, column_idx);
-
-#undef JNI_CHECK_EXTRA
-#define JNI_CHECK_EXTRA
-
-    (*env)->CallVoidMethod(env, cursor, android_database_Cursor_close);
-
-    const char *c_str = (*env)->GetStringUTFChars(env, j_str, 0);
-    JNI_CHECK(c_str, "GetStringUTFChars %p", j_str);
-
-    strncpy(filename, c_str, PATH_MAX);
-    (*env)->ReleaseStringUTFChars(env, j_str, c_str);
-
-    (*env)->PopLocalFrame(env, NULL);
-    return JNI_OK;
-}
-
-int match_ext_from_content(const char *content, const char *extensions) {
-    char filename[PATH_MAX]; // = DocumentFile.fromSingleUri(appContext, contentUri).getName();
-    if (get_filename_from_content(content, filename) != JNI_OK)
-        return 0;
-    LOGE("%s:%d recovered name '%s'", __PRETTY_FUNCTION__, __LINE__, filename);
-    return av_match_ext(filename, extensions);
-}
 /**
  * Sets log level.
  *
